Refactors and cleans translations code
